backend/requirements.txt
autopep8
Django
django-import-export
mysqlclient
django-admin-interface
psycopg2-binary>=2.8
django-environ
django-admin-rangefilter